+++ innopoints/views/__init__.py
@@ -1,1 +1,3 @@
+from .authentication import *
+
 from .views import *  # temporary, until all views have been moved
