+++ innopoints/views/views.py
@@ -4,20 +4,18 @@ import mimetypes
 
 import requests
 import werkzeug
-from authlib.jose.errors import MissingClaimError, InvalidClaimError
-from flask import abort, jsonify, request, current_app, url_for, redirect
+from flask import abort, jsonify, request, current_app
 from flask.views import MethodView
-from flask_login import login_user, login_required, logout_user, current_user
+from flask_login import login_required, current_user
 from marshmallow import ValidationError
 from sqlalchemy import or_
 from sqlalchemy.exc import IntegrityError
 
-from innopoints.extensions import oauth, db
+from innopoints.extensions import db
 from innopoints.blueprints import api
 import innopoints.file_manager_s3 as file_manager
 from innopoints.models import (
     Activity,
-    Account,
     Color,
     Competence,
     LifetimeStage,
@@ -803,73 +801,3 @@ def create_size():
         abort(400, {'message': 'Data integrity violated.'})
 
     return in_out_schema.jsonify(new_size)
-
-
-# ----- Authorization -----
-
-@api.route('/login', methods=['GET'])
-def login():
-    """Redirect the user to the Innopolis SSO login page"""
-    redirect_uri = url_for('api.authorize', _external=True)
-    return oauth.innopolis_sso.authorize_redirect(redirect_uri)
-
-
-@api.route('/authorize')
-def authorize():
-    """Catch the user after the back-redirect and fetch the essential info"""
-    token = oauth.innopolis_sso.authorize_access_token(
-        redirect_uri=url_for('api.authorize', _external=True))
-    try:
-        userinfo = oauth.innopolis_sso.parse_id_token(token)
-    except (MissingClaimError, InvalidClaimError):
-        return abort(401)
-
-    user = Account.query.get(userinfo['email'])
-    if user is None:
-        user = Account(email=userinfo['email'],
-                       full_name=userinfo['commonname'],
-                       university_status=userinfo['role'],
-                       is_admin=current_app.config['IS_ADMIN'](userinfo))
-        db.session.add(user)
-        db.session.commit()
-
-    if user.full_name != userinfo['commonname']:
-        user.full_name = userinfo['commonname']
-
-    if user.university_status != userinfo['role']:
-        user.university_status = userinfo['role']
-
-    if user.is_admin != current_app.config['IS_ADMIN'](userinfo):
-        user.is_admin = current_app.config['IS_ADMIN'](userinfo)
-
-    db.session.commit()
-
-    login_user(user, remember=True)
-
-    return redirect(url_for('api.list_projects'))
-
-
-@api.route('/logout')
-def logout():
-    """Log out the currently signed in user"""
-    logout_user()
-    return '', 204
-
-
-@api.route('/login_cheat')
-def login_cheat():
-    """Bypass OAuth"""
-    # TODO: remove this
-    users = Account.query.all()
-    if not users:
-        user = Account(email='debug@only.com',
-                       full_name='Cheat Account',
-                       university_status='hacker',
-                       is_admin=True)
-        db.session.add(user)
-        db.session.commit()
-    else:
-        user = users[0]
-    login_user(user, remember=True)
-
-    return NO_PAYLOAD
